cleanup function and when removing source edge, we need to update previous_component of our Node
