+++ src/utils/freqFuncs.ts
@@ -1,4 +1,5 @@
-import { ContentType } from "../types/Component";
+import { ContentType, ComponentType } from "../types/Component";
+import { Node, XYPosition } from "reactflow";
 
 export const getPathOfContent = (
   id: number,
@@ -7,3 +8,37 @@ export const getPathOfContent = (
   const content = contentTypes?.find((content) => content.id === id);
   return content?.path?.split(".ir")[1];
 };
+
+export function makeNode(
+  data: ComponentType,
+  position: XYPosition,
+): Node<ComponentType> {
+  const {
+    id,
+    previous_component,
+    component_name,
+    component_type,
+    component_content_type,
+    position_x,
+    position_y,
+  } = data;
+
+  const componentData: ComponentType = {
+    id,
+    previous_component,
+    component_name,
+    component_content_type,
+    component_type,
+    position_x,
+    position_y,
+  };
+  const newNode: Node<ComponentType> = {
+    id: `${componentData.id}`,
+    type: "customNode",
+    position: position,
+    selected: false,
+    data: componentData,
+  };
+
+  return newNode;
+}
